updated for version 7.0063

+++ src/installml.sh
@@ -38,32 +38,32 @@ rgviewname=${14}
 gvimdiffname=${15}
 eviewname=${16}
 
-if test $what = "install"; then
+if test $what = "install" -a (-e $destdir/$vimname.1 -o -e $destdir/$vimdiffname.1 -o -e $destdir/$eviewname.1); then
    if test ! -d $destdir; then
       echo creating $destdir
       ./mkinstalldirs $destdir
    fi
 
    # ex
-   if test ! -e $destdir/$exname.1; then
+   if test ! -e $destdir/$exname.1 -a -e $destdir/$vimname.1; then
       echo creating link $destdir/$exname.1
       cd $destdir; ln -s $vimname.1 $exname.1
    fi
 
    # view
-   if test ! -e $destdir/$viewname.1; then
+   if test ! -e $destdir/$viewname.1 -a -e $destdir/$vimname.1; then
       echo creating link $destdir/$viewname.1
       cd $destdir; ln -s $vimname.1 $viewname.1
    fi
 
    # rvim
-   if test ! -e $destdir/$rvimname.1; then
+   if test ! -e $destdir/$rvimname.1 -a -e $destdir/$vimname.1; then
       echo creating link $destdir/$rvimname.1
       cd $destdir; ln -s $vimname.1 $rvimname.1
    fi
 
    # rview
-   if test ! -e $destdir/$rviewname.1; then
+   if test ! -e $destdir/$rviewname.1 -a -e $destdir/$vimname.1; then
       echo creating link $destdir/$rviewname.1
       cd $destdir; ln -s $vimname.1 $rviewname.1
    fi
@@ -71,37 +71,37 @@ if test $what = "install"; then
    # GUI targets are optional
    if test "$gui" = "yes"; then
       # gvim
-      if test ! -e $destdir/$gvimname.1; then
+      if test ! -e $destdir/$gvimname.1 -a -e $destdir/$vimname.1; then
 	 echo creating link $destdir/$gvimname.1
 	 cd $destdir; ln -s $vimname.1 $gvimname.1
       fi
 
       # gview
-      if test ! -e $destdir/$gviewname.1; then
+      if test ! -e $destdir/$gviewname.1 -a -e $destdir/$vimname.1; then
 	 echo creating link $destdir/$gviewname.1
 	 cd $destdir; ln -s $vimname.1 $gviewname.1
       fi
 
       # rgvim
-      if test ! -e $destdir/$rgvimname.1; then
+      if test ! -e $destdir/$rgvimname.1 -a -e $destdir/$vimname.1; then
 	 echo creating link $destdir/$rgvimname.1
 	 cd $destdir; ln -s $vimname.1 $rgvimname.1
       fi
 
       # rgview
-      if test ! -e $destdir/$rgviewname.1; then
+      if test ! -e $destdir/$rgviewname.1 -a -e $destdir/$vimname.1; then
 	 echo creating link $destdir/$rgviewname.1
 	 cd $destdir; ln -s $vimname.1 $rgviewname.1
       fi
 
       # gvimdiff
-      if test ! -e $destdir/$gvimdiffname.1; then
+      if test ! -e $destdir/$gvimdiffname.1 -a -e $destdir/$vimdiffname.1; then
 	 echo creating link $destdir/$gvimdiffname.1
 	 cd $destdir; ln -s $vimdiffname.1 $gvimdiffname.1
       fi
 
       # eview
-      if test ! -e $destdir/$eviewname.1; then
+      if test ! -e $destdir/$eviewname.1 -a -e $destdir/$evimname.1; then
 	 echo creating link $destdir/$eviewname.1
 	 cd $destdir; ln -s $evimname.1 $eviewname.1
       fi
